Extract JsxAdapter base class for shared SSR logic
Extract duplicated logic from HonoAdapter and TestAdapter into a shared
JsxAdapter base class (extends BaseAdapter):

- generateSignalInitializers() — SSR no-op generation with preserveTypes flag
- formatImportSpecifiers() — import statement formatting
- renderNodeRaw() — raw node rendering for ternary expressions
- renderScopeMarker/SlotMarker/CondMarker() — hydration markers

HonoAdapter and TestAdapter now extend JsxAdapter, reducing duplication
and making it easier to add new JSX-based adapters.
